Moved google py code into seperate file
